+++ src/utils/changeTheme.js
@@ -18,7 +18,7 @@ export default async function changeTheme(selector, stage) {
 
   button.addEventListener('change', () => toggleTheme(stage));
 
-  applyTheme();
+  // applyTheme();
 }
 
 async function toggleTheme(stage) {
@@ -42,7 +42,10 @@ async function applyTheme() {
 }
 
 function updatePageForTheme(theme, stage) {
-  if (window.location.pathname === PORTFOLIO_PATH && stage) {
+  if (
+    (window.location.pathname === PORTFOLIO_PATH && stage) ||
+    (window.location.pathname === '/' && stage)
+  ) {
     updatePortfolioPage(theme, stage);
   }
 }